Moved read more modal to separate file Moved inline styles to stylesheet
